feat: introduce hybrid architecture with WebSocket and API integration
- Implement optimized hybrid architecture combining API REST for initial data and WebSocket for real-time updates
- Enhance SDK with new utility functions for KYC and operation status labels
- Consolidate services into a single IbexClient for improved performance and reduced API calls
- Update documentation to reflect new architecture and usage examples
- Fix various issues related to WebAuthn and operation history display
- Significantly reduce API request count from ~10 to 1 at startup, improving efficiency
