+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,155 @@
+name: Bug Report
+description: Signaler un bug ou un comportement inattendu
+title: '[BUG] '
+labels: ['bug', 'needs-triage']
+assignees: ['AbsconseOfficiel']
+
+body:
+  - type: markdown
+    attributes:
+      value: |
+        Merci de prendre le temps de signaler ce bug !
+
+        **Avant de créer cette issue, vérifiez :**
+        - [ ] Le bug n'a pas déjà été signalé
+        - [ ] Vous utilisez la dernière version du SDK
+        - [ ] Vous avez consulté la [documentation](https://github.com/AbsconseOfficiel/ibex-sdk/tree/main/docs)
+
+  - type: textarea
+    id: description
+    attributes:
+      label: Description du bug
+      description: Décrivez clairement le problème rencontré
+      placeholder: 'Le SDK ne se connecte pas au WebSocket...'
+    validations:
+      required: true
+
+  - type: textarea
+    id: steps
+    attributes:
+      label: Étapes pour reproduire
+      description: Décrivez les étapes pour reproduire le problème
+      placeholder: |
+        1. Installer le SDK
+        2. Configurer avec baseURL
+        3. Appeler useIbex()
+        4. Voir l'erreur...
+    validations:
+      required: true
+
+  - type: textarea
+    id: expected
+    attributes:
+      label: Comportement attendu
+      description: Que devrait-il se passer ?
+      placeholder: 'Le WebSocket devrait se connecter automatiquement...'
+
+  - type: textarea
+    id: actual
+    attributes:
+      label: Comportement observé
+      description: Que se passe-t-il réellement ?
+      placeholder: 'Le WebSocket ne se connecte pas et affiche une erreur...'
+
+  - type: textarea
+    id: logs
+    attributes:
+      label: Logs d'erreur
+      description: Collez ici les logs d'erreur de la console
+      placeholder: |
+        ```
+        [2025-09-30T12:41:35.238Z] ❌ ERROR [HOOK] Erreur de connexion
+        TypeError: The provided value is not of type '(ArrayBuffer or ArrayBufferView)'
+        ```
+
+  - type: textarea
+    id: version
+    attributes:
+      label: Version du SDK
+      description: Quelle version du SDK utilisez-vous ?
+      options:
+        - '1.1.1'
+        - '1.1.0'
+        - '1.0.0'
+        - 'Autre (préciser dans les commentaires)'
+    validations:
+      required: true
+
+  - type: dropdown
+    id: environment
+    attributes:
+      label: Environnement
+      description: Dans quel environnement le bug se produit-il ?
+      options:
+        - 'Développement local'
+        - 'Staging'
+        - 'Production'
+        - 'Test'
+    validations:
+      required: true
+
+  - type: input
+    id: browser
+    attributes:
+      label: Navigateur
+      description: Quel navigateur utilisez-vous ?
+      placeholder: 'Chrome 118, Firefox 119, Safari 17, etc.'
+
+  - type: input
+    id: os
+    attributes:
+      label: Système d'exploitation
+      description: Sur quel OS ?
+      placeholder: 'Windows 11, macOS 14, Ubuntu 22.04, etc.'
+
+  - type: textarea
+    id: config
+    attributes:
+      label: Configuration
+      description: Partagez votre configuration (sans données sensibles)
+      placeholder: |
+        ```typescript
+        const config = {
+          baseURL: 'https://api.ibexwallet.org',
+          domain: 'example.com',
+        };
+        ```
+
+  - type: textarea
+    id: code
+    attributes:
+      label: Code minimal
+      description: Code minimal pour reproduire le bug
+      placeholder: |
+        ```typescript
+        import { useIbex } from '@absconse/ibex-sdk';
+
+        function App() {
+          const { signIn, error } = useIbex(config);
+          // ...
+        }
+        ```
+
+  - type: checkboxes
+    id: checks
+    attributes:
+      label: Vérifications
+      description: Cochez les éléments qui s'appliquent
+      options:
+        - label: "J'ai vérifié que le bug n'existe pas déjà"
+          required: true
+        - label: "J'ai testé avec la dernière version"
+          required: true
+        - label: "J'ai consulté la documentation"
+          required: true
+        - label: "J'ai fourni toutes les informations demandées"
+          required: true
+
+  - type: markdown
+    attributes:
+      value: |
+        ## Informations supplémentaires
+
+        Si vous avez des captures d'écran, des fichiers de log, ou d'autres informations utiles, ajoutez-les dans les commentaires après avoir créé l'issue.
+
+        **Merci de votre contribution !**
