+++ .github/pull_request_template.md
@@ -1,0 +1,130 @@
+# Pull Request
+
+## Description
+
+<!-- Décrivez clairement les changements apportés -->
+
+## Issue liée
+
+<!-- Lien vers l'issue : Fixes #123 -->
+
+## Type de changement
+
+- [ ] Bug fix (changement non-breaking qui corrige un problème)
+- [ ] Nouvelle fonctionnalité (changement non-breaking qui ajoute une fonctionnalité)
+- [ ] Breaking change (changement qui pourrait casser la compatibilité)
+- [ ] Documentation (changement uniquement dans la documentation)
+- [ ] Refactoring (changement de code qui ne corrige pas de bug ni n'ajoute de fonctionnalité)
+- [ ] Performance (changement qui améliore les performances)
+- [ ] Test (ajout ou modification de tests)
+- [ ] Build (changement dans le système de build ou les dépendances)
+
+## Tests
+
+### Tests effectués
+
+- [ ] Tests unitaires passent
+- [ ] Tests d'intégration passent
+- [ ] Tests manuels effectués
+- [ ] Aucune régression détectée
+
+### Instructions de test
+
+<!-- Décrivez comment tester vos changements -->
+
+```bash
+# Commandes pour tester
+npm run test
+npm run build
+npm run lint
+```
+
+## Captures d'écran (si applicable)
+
+<!-- Ajoutez des captures d'écran pour les changements visuels -->
+
+## Checklist
+
+### Code
+
+- [ ] Mon code suit les conventions du projet
+- [ ] J'ai effectué une auto-review de mon code
+- [ ] J'ai commenté les parties complexes de mon code
+- [ ] Mes changements ne génèrent pas de nouveaux warnings
+- [ ] J'ai ajouté des tests qui prouvent que mon fix est efficace
+- [ ] Les tests existants passent avec mes changements
+- [ ] J'ai mis à jour la documentation si nécessaire
+
+### Documentation
+
+- [ ] J'ai mis à jour la documentation si nécessaire
+- [ ] J'ai mis à jour le CHANGELOG.md
+- [ ] J'ai mis à jour les types TypeScript si nécessaire
+- [ ] J'ai vérifié que tous les exemples fonctionnent
+
+### Sécurité
+
+- [ ] Mes changements n'introduisent pas de vulnérabilités de sécurité
+- [ ] J'ai validé toutes les entrées utilisateur
+- [ ] J'ai utilisé des types stricts TypeScript
+- [ ] Je n'ai pas committé de données sensibles
+
+## Déploiement
+
+### Impact sur le déploiement
+
+- [ ] Aucun impact sur le déploiement
+- [ ] Impact mineur (ajout de fonctionnalités)
+- [ ] Impact majeur (changements breaking)
+
+### Actions requises
+
+- [ ] Aucune action requise
+- [ ] Mise à jour de la documentation
+- [ ] Communication aux utilisateurs
+- [ ] Migration des données
+
+## Métriques
+
+### Performance
+
+- [ ] Aucun impact sur les performances
+- [ ] Amélioration des performances
+- [ ] Impact négatif sur les performances (justifié)
+
+### Bundle size
+
+- [ ] Aucun impact sur la taille du bundle
+- [ ] Réduction de la taille du bundle
+- [ ] Augmentation de la taille du bundle (justifiée)
+
+## Review
+
+### Points d'attention pour les reviewers
+
+<!-- Mentionnez les points spécifiques à vérifier -->
+
+### Questions pour les reviewers
+
+<!-- Posez des questions spécifiques si nécessaire -->
+
+## Ressources
+
+### Documentation mise à jour
+
+- [ ] README.md
+- [ ] docs/hooks.md
+- [ ] docs/api-reference.md
+- [ ] docs/examples.md
+- [ ] docs/types.md
+- [ ] Autre : \***\*\_\_\_\*\***
+
+### Liens utiles
+
+- [Documentation](https://github.com/AbsconseOfficiel/ibex-sdk/tree/main/docs)
+- [Guide de contribution](https://github.com/AbsconseOfficiel/ibex-sdk/blob/main/CONTRIBUTING.md)
+- [Code de conduite](https://github.com/AbsconseOfficiel/ibex-sdk/blob/main/CODE_OF_CONDUCT.md)
+
+---
+
+**Merci pour votre contribution !**
