+++ .npmignore
@@ -54,3 +54,4 @@ logs/
 # Temporary files
 *.tmp
 *.temp
+.cursor
