+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,89 @@
+# Code de conduite
+
+## Notre engagement
+
+Dans l'intérêt de favoriser un environnement ouvert et accueillant, nous nous engageons, en tant que contributeurs et mainteneurs, à faire de la participation à notre projet et à notre communauté une expérience sans harcèlement pour tous, quel que soit l'âge, la taille, le handicap, l'ethnicité, les caractéristiques sexuelles, l'identité et l'expression de genre, le niveau d'expérience, l'éducation, le statut socio-économique, la nationalité, l'apparence personnelle, la religion ou l'orientation sexuelle.
+
+## Nos standards
+
+### Comportements qui contribuent à créer un environnement positif
+
+- **Respect mutuel** : Traiter tout le monde avec respect et dignité
+- **Collaboration constructive** : Être ouvert aux idées et suggestions
+- **Communication claire** : Exprimer ses idées de manière claire et constructive
+- **Empathie** : Comprendre et respecter les points de vue différents
+- **Apprentissage** : Accepter les critiques constructives et apprendre des erreurs
+- **Inclusion** : Accueillir les nouveaux contributeurs et les aider
+
+### Comportements inacceptables
+
+- **Harcèlement** : Commentaires offensants, insultes ou attaques personnelles
+- **Discrimination** : Toute forme de discrimination basée sur des caractéristiques personnelles
+- **Trolling** : Publier des commentaires délibérément provocants ou off-topic
+- **Spam** : Publier du contenu non pertinent ou répétitif
+- **Menaces** : Menaces de violence ou d'intimidation
+- **Contenu inapproprié** : Contenu sexuellement explicite ou violent
+
+## Signalement et sanctions
+
+### Comment signaler
+
+Si vous êtes témoin ou victime d'un comportement inapproprié :
+
+1. **LinkedIn** : [Dylan Enjolvin](https://www.linkedin.com/in/dylanenjolvin/)
+2. **GitHub** : Créer une issue privée avec le label "conduct"
+
+### Processus de traitement
+
+1. **Accusé de réception** : Confirmation dans les 72h
+2. **Investigation** : Analyse de la situation
+3. **Action** : Mesures appropriées
+4. **Suivi** : Vérification de l'efficacité
+
+### Sanctions possibles
+
+- **Avertissement** : Pour les premiers manquements mineurs
+- **Suspension temporaire** : Pour les violations répétées
+- **Bannissement permanent** : Pour les violations graves
+
+## Nos responsabilités
+
+### Responsabilités des mainteneurs
+
+- **Clarifier** les standards de comportement acceptables
+- **Prendre** des mesures correctives appropriées
+- **Supprimer** le contenu inapproprié
+- **Bannir** temporairement ou définitivement les contributeurs
+
+### Responsabilités des contributeurs
+
+- **Respecter** ce code de conduite
+- **Signaler** les comportements inappropriés
+- **Collaborer** de manière constructive
+- **Maintenir** un environnement positif
+
+## Portée
+
+Ce code de conduite s'applique à :
+
+- **Tous les espaces** du projet (GitHub, discussions, issues, PRs)
+- **Tous les événements** publics organisés par le projet
+- **Toutes les communications** liées au projet
+- **Tous les contributeurs** (mainteneurs, contributeurs, utilisateurs)
+
+## Contact
+
+- **LinkedIn** : [Dylan Enjolvin](https://www.linkedin.com/in/dylanenjolvin/)
+- **GitHub** : [@AbsconseOfficiel](https://github.com/AbsconseOfficiel)
+
+## Attribution
+
+Ce code de conduite est adapté du [Contributor Covenant](https://www.contributor-covenant.org/), version 2.1, disponible à [https://www.contributor-covenant.org/version/2/1/code_of_conduct.html](https://www.contributor-covenant.org/version/2/1/code_of_conduct.html).
+
+## Reconnaissance
+
+Nous remercions tous les contributeurs qui respectent ce code de conduite et contribuent à créer une communauté positive et inclusive.
+
+---
+
+_Dernière mise à jour : 30 septembre 2025_
